+++ json_schema_for_humans/template_renderer.py
@@ -1,7 +1,6 @@
 import re
 from typing import Optional
 
-import htmlmin  # type: ignore  # No stubs available
 import jinja2
 import markdown2  # type: ignore  # No stubs available
 from jinja2 import FileSystemLoader, Template
@@ -11,14 +10,6 @@ from json_schema_for_humans import jinja_filters, templating_utils
 from json_schema_for_humans.generation_configuration import GenerationConfiguration
 from json_schema_for_humans.md_template import MarkdownTemplate
 from json_schema_for_humans.schema.schema_node import SchemaNode
-
-
-def _minify(rendered: str, is_markdown: bool, is_html: bool) -> str:
-    if is_markdown:
-        return re.sub(r"\n\s*\n", "\n\n", rendered)
-    if is_html:
-        return htmlmin.minify(rendered)
-    return rendered
 
 
 class TemplateRenderer:
@@ -31,8 +22,8 @@ class TemplateRenderer:
         env = jinja2.Environment(
             loader=loader,
             extensions=[loopcontrols],
-            trim_blocks=(self.config.template_is_markdown),
-            lstrip_blocks=(self.config.template_is_markdown),
+            trim_blocks=self.config.template_is_markdown,
+            lstrip_blocks=self.config.template_is_markdown,
         )
         env.globals["jsfh_config"] = self.config
         env.globals["jsfh_md"] = markdown2.Markdown(extras=self.config.markdown_options)
@@ -73,7 +64,7 @@ class TemplateRenderer:
     def render(self, intermediate_schema: SchemaNode) -> str:
         rendered = self.template.render(schema=intermediate_schema, config=self.config)
 
-        if self.config.minify:
-            rendered = _minify(rendered, self.config.template_is_markdown, self.config.template_is_html)
+        if self.config.template_is_markdown:
+            rendered = re.sub(r"\n\s*\n", "\n\n", rendered)
 
         return rendered